ocr/requirements.txt
ultralytics>=8.1.0
opencv-python>=4.9.0
pdf2image>=1.17.0
numpy>=1.26.0
Pillow>=10.2.0
torch>=2.2.0
transformers>=4.38.0
scikit-learn
